[feat] Point 로직에도 분산락 Proxy 구현
