+++ src/test/java/kr/hhplus/be/server/point/application/PointServiceImplTest.java
@@ -17,7 +17,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-class PointServiceTest {
+class PointServiceImplTest {
 
     @InjectMocks
     private PointService pointService;
